test(ui/weather/widgets): Move and update widget tests with adjusted imports
