+++ test/ui/weather/widgets/city_search_input_test.dart
@@ -6,7 +6,7 @@ import 'package:weather_app/core/strings/city_name_validator_strings.dart';
 import 'package:weather_app/core/strings/city_search_input_strings.dart';
 import 'package:weather_app/ui/weather/widgets/city_search_input.dart';
 
-import '../mocks/mock_firebase.dart';
+import '../../../features/weather/mocks/mock_firebase.dart';
 
 void main() {
   setUpAll(() async {
